```
hyperfine --warmup 1 --runs 5 --export-json me-timing.json ./calculate_average_bytesfellow.sh
Benchmark 1: ./calculate_average_bytesfellow.sh
  Time (mean ± σ):     19.645 s ±  2.991 s    [User: 128.586 s, System: 6.427 s]
  Range (min … max):   16.515 s … 23.829 s    5 runs
```

macbook pro m1 10 cores.
baseline runs in 189 seconds on the same generated input file.